+++ src/core/plugins/teleportEffect/teleporter.ts
@@ -4,9 +4,10 @@ import ChatController from '../../server/systems/chat';
 import { ItemType } from '../../shared/enums/itemType';
 import { Permissions } from '../../shared/flags/permissions';
 import { Item } from '../../shared/interfaces/Item';
+import { LOCALE_KEYS } from '../../shared/locale/languages/keys';
+import { LocaleController } from '../../shared/locale/locale';
 import { deepCloneObject } from '../../shared/utility/deepCopy';
 
-const teleporterDesc = '/teleporter - Adds item for current position.';
 const teleporterItem: Item = {
     name: `Teleporter`,
     uuid: `teleporter`,
@@ -20,17 +21,22 @@ const teleporterItem: Item = {
     }
 };
 
-ChatController.addCommand('teleporter', teleporterDesc, Permissions.Admin, (player: alt.Player) => {
-    let itemClone = deepCloneObject<Item>(teleporterItem);
-    let slotInfo = playerFuncs.inventory.getFreeInventorySlot(player);
-    itemClone.data.x = player.pos.x;
-    itemClone.data.y = player.pos.y;
-    itemClone.data.z = player.pos.z;
+ChatController.addCommand(
+    'teleporter',
+    LocaleController.get(LOCALE_KEYS.COMMAND_TELEPORTER, '/teleporter'),
+    Permissions.Admin,
+    (player: alt.Player) => {
+        let itemClone = deepCloneObject<Item>(teleporterItem);
+        let slotInfo = playerFuncs.inventory.getFreeInventorySlot(player);
+        itemClone.data.x = player.pos.x;
+        itemClone.data.y = player.pos.y;
+        itemClone.data.z = player.pos.z;
 
-    playerFuncs.inventory.inventoryAdd(player, itemClone, slotInfo.slot, slotInfo.tab);
-    playerFuncs.save.field(player, 'inventory', player.data.inventory);
-    playerFuncs.sync.inventory(player);
-});
+        playerFuncs.inventory.inventoryAdd(player, itemClone, slotInfo.slot, slotInfo.tab);
+        playerFuncs.save.field(player, 'inventory', player.data.inventory);
+        playerFuncs.sync.inventory(player);
+    }
+);
 
 alt.on('effect:Teleport', (player: alt.Player, item: Item) => {
     if (!item || !item.data || !item.data.x || !item.data.y || !item.data.z) {