+++ src/core/server/commands/job.ts
@@ -1,10 +1,17 @@
 import * as alt from 'alt-server';
 import { Permissions } from '../../shared/flags/permissions';
+import { LOCALE_KEYS } from '../../shared/locale/languages/keys';
+import { LocaleController } from '../../shared/locale/locale';
 import { playerFuncs } from '../extensions/Player';
 import ChatController from '../systems/chat';
 import { getPlayerJob } from '../systems/job';
 
-ChatController.addCommand('quitjob', '/quitjob - Will quit your current job.', Permissions.None, handleCommand);
+ChatController.addCommand(
+    'quitjob',
+    LocaleController.get(LOCALE_KEYS.COMMAND_QUIT_JOB, '/quitjob'),
+    Permissions.None,
+    handleCommand
+);
 
 function handleCommand(player: alt.Player): void {
     if (!player || !player.valid) {