+++ src/core/server/commands/setFood.ts
@@ -1,9 +1,16 @@
 import * as alt from 'alt-server';
 import { Permissions } from '../../shared/flags/permissions';
+import { LOCALE_KEYS } from '../../shared/locale/languages/keys';
+import { LocaleController } from '../../shared/locale/locale';
 import { playerFuncs } from '../extensions/Player';
 import ChatController from '../systems/chat';
 
-ChatController.addCommand('setfood', '/setfood [amount] - Set the food level', Permissions.Admin, handleCommand);
+ChatController.addCommand(
+    'setfood',
+    LocaleController.get(LOCALE_KEYS.COMMAND_SET_FOOD, '/setfood'),
+    Permissions.Admin,
+    handleCommand
+);
 
 function handleCommand(player: alt.Player, value: string): void {
     let foodLevel = parseInt(value);