+++ src/core/server/commands/setWater.ts
@@ -1,9 +1,16 @@
 import * as alt from 'alt-server';
 import { Permissions } from '../../shared/flags/permissions';
+import { LOCALE_KEYS } from '../../shared/locale/languages/keys';
+import { LocaleController } from '../../shared/locale/locale';
 import { playerFuncs } from '../extensions/Player';
 import ChatController from '../systems/chat';
 
-ChatController.addCommand('setwater', '/setwater [amount] - Set player water level', Permissions.Admin, handleCommand);
+ChatController.addCommand(
+    'setwater',
+    LocaleController.get(LOCALE_KEYS.COMMAND_SET_WATER, '/setwater'),
+    Permissions.Admin,
+    handleCommand
+);
 
 function handleCommand(player: alt.Player, value: string): void {
     let waterLevel = parseInt(value);