+++ documentation/changelog.md
@@ -35,6 +35,7 @@ description: List of all the changes and current versions.
 + Weapons only removed when a player is sent to the Hospital
 + Added New Locale System
 + New Locale System Supports Multiple Languages (Only English thus far)
++ Added Locale for All Commands
 ```
 
 ## 1.8.0 - February 28-x, 2021